Disallow leading zeros
Leading zeros, that is zeros, which occur sequentially at the beginning of the number component, prior to a non-zero digit, are no longer allowed.

This is in line with semver. Since this library is used in places where only the version core is allowed, and the shorthand version of the version core is allowed, it is useful to otherwise follow the semver spec.